requirements.txt
Django==4.2.11
djangorestframework==3.15.1
dj-rest-auth==5.1.0
drf-yasg==1.21.7
psycopg2==2.9.9
django-filter==24.2
